Split Gaussian 2D integrator into separate file
Also update some variable names
