+++ src/gaussian.h
@@ -36,18 +36,6 @@ typedef py::array_t<double> paramsgauss;
 namespace multiprofit {
 
 /*
-    Numerically integrate a 2D Gaussian centered on (XCEN, YCEN) with magnitude MAG, half-light radius RE,
-    position angle ANG (annoying GALFIT convention of up=0), axis ratio AXRAT, over a grid of defined by the
-    corners (XMIN, YMIN) and (XMAX, YMAX) with XDIM x YDIM pixels, to a relative tolerance ACC.
-*/
-ndarray make_gaussian(
-    const double XCEN, const double YCEN, const double MAG,
-    const double RE, const double AXRAT, const double ANG,
-    const double XMIN, const double XMAX, const double YMIN, const double YMAX,
-    const unsigned int XDIM, const unsigned int YDIM,
-    const double ACC);
-
-/*
     Efficiently evaluate a 2D Gaussian centered on (XCEN, YCEN) with total flux L, half-light radius R,
     position angle ANG (annoying GALFIT convention of up=0), axis ratio AXRAT, at the centers of pixels on a
     grid defined by the corners (XMIN, YMIN) and (XMAX, YMAX) with XDIM x YDIM pixels.
